refactor(sam3): specialize family implementation
Consolidate sam3 model construction and weight mapping under its owned package. Remove unused family-local duplicates while preserving the existing plugin API.
