+++ python/tensorrt_model_connect/families/sam3/MODEL.toml
@@ -1,0 +1,5 @@
+id = "sam3"
+plugin = "sam3"
+module = "plugin"
+aliases = ["sam3"]
+prefixes = ["sam3"]
