+++ python/tensorrt_model_connect/families/sam3/__init__.py
@@ -1,5 +1,3 @@
-"""SAM3 family package."""
-
 from .plugin import plugin
 
 __all__ = ["plugin"]
