+++ python/tensorrt_model_connect/families/sam3/model/__init__.py
@@ -1,0 +1,1 @@
+"""Family-owned TensorRT model construction components."""
