+++ python/tensorrt_model_connect/families/sam3/model/components/text_encoder.py
@@ -14,10 +14,10 @@ import numpy as np
 
 from tensorrt_model_connect import trt_compat
 
-from . import graph_ops
+from .. import model as graph_ops
 
 if TYPE_CHECKING:
-    from .checkpoint_mapper import WeightDict
+    from ...weights import WeightDict
 
 
 def _trt():
@@ -45,8 +45,7 @@ def build_sam3_text_encoder_engine(
 
     logger = trt.Logger(trt.Logger.VERBOSE if verbose else trt.Logger.WARNING)
     builder = trt.Builder(logger)
-    network = builder.create_network(
-        1 << int(trt.NetworkDefinitionCreationFlag.STRONGLY_TYPED))
+    network = builder.create_network(1 << int(trt.NetworkDefinitionCreationFlag.STRONGLY_TYPED))
     config = builder.create_builder_config()
     config.set_memory_pool_limit(trt.MemoryPoolType.WORKSPACE, 4 << 30)
 
@@ -62,30 +61,36 @@ def build_sam3_text_encoder_engine(
 
     pos_ids_np = np.arange(max_seq_len, dtype=np.int32)
     pos_ids = network.add_constant(
-        (max_seq_len,), trt.Weights(np.ascontiguousarray(pos_ids_np))).get_output(0)
+        (max_seq_len,), trt.Weights(np.ascontiguousarray(pos_ids_np))
+    ).get_output(0)
     pos_const = graph_ops.add_constant(network, pos_embed_w.shape, pos_embed_w)
     pos_gather = network.add_gather(pos_const, pos_ids, axis=0)
     hidden = network.add_elementwise(
-        hidden, pos_gather.get_output(0), trt.ElementWiseOperation.SUM).get_output(0)
+        hidden, pos_gather.get_output(0), trt.ElementWiseOperation.SUM
+    ).get_output(0)
 
     causal_np = np.full((max_seq_len, max_seq_len), -1e9, dtype=np.float32)
     for i in range(max_seq_len):
         causal_np[i, : i + 1] = 0.0
     causal_const = graph_ops.add_constant(
-        network, (1, 1, max_seq_len, max_seq_len), causal_np[None, None])
+        network, (1, 1, max_seq_len, max_seq_len), causal_np[None, None]
+    )
     mask_float = network.add_cast(attention_mask, trt.float32).get_output(0)
     valid_ones = graph_ops.add_constant(
-        network, (max_seq_len,), np.ones((max_seq_len,), dtype=np.float32))
+        network, (max_seq_len,), np.ones((max_seq_len,), dtype=np.float32)
+    )
     invalid_mask = network.add_elementwise(
-        valid_ones, mask_float, trt.ElementWiseOperation.SUB).get_output(0)
+        valid_ones, mask_float, trt.ElementWiseOperation.SUB
+    ).get_output(0)
     invalid_mask_4d = network.add_shuffle(invalid_mask)
     invalid_mask_4d.reshape_dims = (1, 1, 1, max_seq_len)
-    mask_penalty = graph_ops.add_constant(
-        network, (1, 1, 1, 1), np.array([-1e9], dtype=np.float32))
+    mask_penalty = graph_ops.add_constant(network, (1, 1, 1, 1), np.array([-1e9], dtype=np.float32))
     padding_bias = network.add_elementwise(
-        invalid_mask_4d.get_output(0), mask_penalty, trt.ElementWiseOperation.PROD)
+        invalid_mask_4d.get_output(0), mask_penalty, trt.ElementWiseOperation.PROD
+    )
     attention_bias = network.add_elementwise(
-        causal_const, padding_bias.get_output(0), trt.ElementWiseOperation.SUM).get_output(0)
+        causal_const, padding_bias.get_output(0), trt.ElementWiseOperation.SUM
+    ).get_output(0)
 
     for layer_idx in range(num_layers):
         prefix = f"text_model.encoder.layers.{layer_idx}"
@@ -100,20 +105,23 @@ def build_sam3_text_encoder_engine(
         )
 
         q = graph_ops.add_matmul_rhs_constant(
-            network, normed, hidden_size, hidden_size,
-            weights[f"{prefix}.self_attn.q_proj.weight"])
+            network, normed, hidden_size, hidden_size, weights[f"{prefix}.self_attn.q_proj.weight"]
+        )
         q = graph_ops.add_bias_sum(
-            network, q, hidden_size, weights[f"{prefix}.self_attn.q_proj.bias"])
+            network, q, hidden_size, weights[f"{prefix}.self_attn.q_proj.bias"]
+        )
         k = graph_ops.add_matmul_rhs_constant(
-            network, normed, hidden_size, hidden_size,
-            weights[f"{prefix}.self_attn.k_proj.weight"])
+            network, normed, hidden_size, hidden_size, weights[f"{prefix}.self_attn.k_proj.weight"]
+        )
         k = graph_ops.add_bias_sum(
-            network, k, hidden_size, weights[f"{prefix}.self_attn.k_proj.bias"])
+            network, k, hidden_size, weights[f"{prefix}.self_attn.k_proj.bias"]
+        )
         v = graph_ops.add_matmul_rhs_constant(
-            network, normed, hidden_size, hidden_size,
-            weights[f"{prefix}.self_attn.v_proj.weight"])
+            network, normed, hidden_size, hidden_size, weights[f"{prefix}.self_attn.v_proj.weight"]
+        )
         v = graph_ops.add_bias_sum(
-            network, v, hidden_size, weights[f"{prefix}.self_attn.v_proj.bias"])
+            network, v, hidden_size, weights[f"{prefix}.self_attn.v_proj.bias"]
+        )
 
         context = graph_ops.add_attention_from_rows(
             network,
@@ -128,13 +136,18 @@ def build_sam3_text_encoder_engine(
         )
 
         attn_out = graph_ops.add_matmul_rhs_constant(
-            network, context, hidden_size, hidden_size,
-            weights[f"{prefix}.self_attn.out_proj.weight"])
+            network,
+            context,
+            hidden_size,
+            hidden_size,
+            weights[f"{prefix}.self_attn.out_proj.weight"],
+        )
         attn_out = graph_ops.add_bias_sum(
-            network, attn_out, hidden_size,
-            weights[f"{prefix}.self_attn.out_proj.bias"])
-        hidden = network.add_elementwise(
-            hidden, attn_out, trt.ElementWiseOperation.SUM).get_output(0)
+            network, attn_out, hidden_size, weights[f"{prefix}.self_attn.out_proj.bias"]
+        )
+        hidden = network.add_elementwise(hidden, attn_out, trt.ElementWiseOperation.SUM).get_output(
+            0
+        )
 
         normed2 = graph_ops.add_layer_norm_native(
             network,
@@ -145,27 +158,26 @@ def build_sam3_text_encoder_engine(
             eps,
         )
         fc1 = graph_ops.add_matmul_rhs_constant(
-            network, normed2, hidden_size, intermediate_size,
-            weights[f"{prefix}.mlp.fc1.weight"])
+            network, normed2, hidden_size, intermediate_size, weights[f"{prefix}.mlp.fc1.weight"]
+        )
         fc1 = graph_ops.add_bias_sum(
-            network, fc1, intermediate_size, weights[f"{prefix}.mlp.fc1.bias"])
+            network, fc1, intermediate_size, weights[f"{prefix}.mlp.fc1.bias"]
+        )
         if hidden_act == "quick_gelu":
-            coeff = graph_ops.add_constant(
-                network, (1, 1), np.array([1.702], dtype=np.float32))
+            coeff = graph_ops.add_constant(network, (1, 1), np.array([1.702], dtype=np.float32))
             scaled = network.add_elementwise(fc1, coeff, trt.ElementWiseOperation.PROD)
             sigmoid = network.add_activation(scaled.get_output(0), trt.ActivationType.SIGMOID)
             activated = network.add_elementwise(
-                fc1, sigmoid.get_output(0), trt.ElementWiseOperation.PROD).get_output(0)
+                fc1, sigmoid.get_output(0), trt.ElementWiseOperation.PROD
+            ).get_output(0)
         else:
             activated = graph_ops.add_gelu_new(network, fc1)
 
         fc2 = graph_ops.add_matmul_rhs_constant(
-            network, activated, intermediate_size, hidden_size,
-            weights[f"{prefix}.mlp.fc2.weight"])
-        fc2 = graph_ops.add_bias_sum(
-            network, fc2, hidden_size, weights[f"{prefix}.mlp.fc2.bias"])
-        hidden = network.add_elementwise(
-            hidden, fc2, trt.ElementWiseOperation.SUM).get_output(0)
+            network, activated, intermediate_size, hidden_size, weights[f"{prefix}.mlp.fc2.weight"]
+        )
+        fc2 = graph_ops.add_bias_sum(network, fc2, hidden_size, weights[f"{prefix}.mlp.fc2.bias"])
+        hidden = network.add_elementwise(hidden, fc2, trt.ElementWiseOperation.SUM).get_output(0)
 
     hidden = graph_ops.add_layer_norm_native(
         network,
@@ -185,7 +197,8 @@ def build_sam3_text_encoder_engine(
     )
     if "text_projection.bias" in weights:
         projection = graph_ops.add_bias_sum(
-            network, projection, projected_size, weights["text_projection.bias"])
+            network, projection, projected_size, weights["text_projection.bias"]
+        )
 
     projected = network.add_cast(projection, trt.float32).get_output(0)
     projected.name = "sam3_text_features"
